Swizzle api card list
