+++ src/theme/DocCardList/index.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import clsx from 'clsx';
+import {
+  useCurrentSidebarCategory,
+  filterDocCardListItems,
+} from '@docusaurus/theme-common';
+import DocCard from '@theme/DocCard';
+import {ApiDocCard} from "./ApiDocCard";
+function DocCardListForCurrentSidebarCategory({className}) {
+  const category = useCurrentSidebarCategory();
+  return <DocCardList items={category.items} className={className} />;
+}
+export default function DocCardList(props) {
+  const {items, className} = props;
+  if (!items) {
+    return <DocCardListForCurrentSidebarCategory {...props} />;
+  }
+  const filteredItems = filterDocCardListItems(items);
+  return (
+    <section className={clsx('row', className)}>
+      {filteredItems.map((item, index) => (
+        <article key={index} className={ "col col--6 margin-bottom--lg"}>
+          {
+            item.className.includes("api-method")
+                ? <ApiDocCard item={item} />
+                : <DocCard item={item} />
+          }
+        </article>
+      ))}
+    </section>
+  );
+}
